Multinotify 3.0 prerelease
Added multinotify 3.0 prerelease that supports Google Home devices and all other that can handle standard TTS.
There is an actual limit where Google Home don't resume music after TTS and, if not playing music, don't know what's the actual volume of device. This will be worked in the future as there are some limitations of the platform.